Build pocketcoin-cli without server dependencies
