+++ src/rpc/request.h
@@ -7,6 +7,7 @@
 #define POCKETCOIN_RPC_REQUEST_H
 
 #include "pocketdb/SQLiteConnection.h"
+#include "rpc/requestutils.h"
 #include <string>
 
 #include <univalue.h>
@@ -14,20 +15,6 @@
 namespace util {
 class Ref;
 } // namespace util
-
-UniValue JSONRPCRequestObj(const std::string& strMethod, const UniValue& params, const UniValue& id);
-UniValue JSONRPCReplyObj(const UniValue& result, const UniValue& error, const UniValue& id);
-std::string JSONRPCReply(const UniValue& result, const UniValue& error, const UniValue& id);
-UniValue JSONRPCError(int code, const std::string& message);
-
-/** Generate a new RPC authentication cookie and write it to disk */
-bool GenerateAuthCookie(std::string *cookie_out);
-/** Read the RPC authentication cookie from disk */
-bool GetAuthCookie(std::string *cookie_out);
-/** Delete RPC authentication cookie from disk */
-void DeleteAuthCookie();
-/** Parse JSON-RPC batch reply into a vector */
-std::vector<UniValue> JSONRPCProcessBatchReply(const UniValue& in);
 
 class JSONRPCRequest
 {
